requirements.txt
pandas
numpy<2
tensorflow==2.10.1
tensorflow_text==2.10.1
keras==2.10.0
keras-nlp==0.6.0
sklearn.preprocessing
scikit-learn
pyarrow
fastparquet
huggingface-hub
dill
flask
-e .